Add employee function implementedAdd employee function implemented
The roles that correspond to the employee controller have been created.
Within the create function, what is necessary was implemented to insert a new employee into the database.
The view was also created with the form that corresponds to a new user

These changes were implemented to be able to insert a new employee into the database

Co-authored-by: RobertoCarMed <betocarrasco3652@gmail.com>
Co-authored-by: mrivh <m.alicia.riv@gmail.com>