Add table reports
A table called reports was added with the check in and check out fields.
This table will serve to keep a record of attendance and to generate attendance reports.
The form was also modified to add and edit employees removing the field assistances. Since the assists will be handled in a single table, it is not necessary for this field to be in the form.

This task was carried out in order to start working with the assistance functionality.